+++ demo/runtime/src/runtime.rs
@@ -21,7 +21,8 @@ use codec::Slicable;
 use runtime_support::Hashable;
 use runtime_io::{enumerated_trie_root, storage_root, blake2_256};
 use {block, demo_primitives, transaction};
-use {runtime_primitives, consensus, timestamp, system, session, staking, democracy, executive};
+use runtime_primitives;
+use {consensus, council, democracy, executive, session, staking, system, timestamp};
 use runtime_primitives::{Identity, HasPublicAux};
 
 // TODO: move into runtime support/io.
@@ -92,13 +93,16 @@ impl democracy::Trait for Concrete {
 }
 pub type Democracy = democracy::Module<Concrete>;
 
+impl council::Trait for Concrete {}
+pub type Council = council::Module<Concrete>;
+
 impl_outer_dispatch! {
 	pub enum Call where aux: <Concrete as HasPublicAux>::PublicAux {
 		Session = 1,
 		Staking = 2,
 		Timestamp = 3,
 		Democracy = 5,
-//		Council = 6,
+		Council = 6,
 //		CouncilVote = 7,
 	}
 
@@ -107,7 +111,7 @@ impl_outer_dispatch! {
 		Session = 1,
 		Staking = 2,
 		Democracy = 5,
-//		Council = 6,
+		Council = 6,
 //		CouncilVote = 7,
 	}
 }
@@ -117,5 +121,5 @@ pub type Executive = executive::Executive<
 	transaction::CheckedTransaction,
 	Concrete,
 	block::Block,
-	((Democracy, Staking), Session),
+	((Council, Democracy), (Staking, Session)),
 >;
