+++ substrate/runtime/council/Cargo.toml
@@ -14,6 +14,7 @@ substrate-runtime-std = { path = "../../runtime-std", default_features = false }
 substrate-runtime-io = { path = "../../runtime-io", default_features = false }
 substrate-runtime-support = { path = "../../runtime-support", default_features = false }
 substrate-runtime-primitives = { path = "../primitives", default_features = false }
+substrate-runtime-democracy = { path = "../democracy", default_features = false }
 substrate-runtime-session = { path = "../session", default_features = false }
 substrate-runtime-staking = { path = "../staking", default_features = false }
 substrate-runtime-system = { path = "../system", default_features = false }
@@ -29,6 +30,7 @@ std = [
 	"substrate-runtime-io/std",
 	"substrate-runtime-support/std",
 	"substrate-runtime-primitives/std",
+	"substrate-runtime-democracy/std",
 	"substrate-runtime-session/std",
 	"substrate-runtime-staking/std",
 	"substrate-runtime-system/std",
